Image full screen, information, labels are functional.
After tapping on a photo, it will appear in full screen.  The image now retains the aspect ratio of the width and height.
The full screen has a FAB button on the right bottom.  From here, you can access image information, change the labels of the image, and delete the picture.
The information contains the date, locations and labels.
The label modification is a simple addition or removal from the label list.
The delete feature are implemented but not perfect. After deleting the image, there is still a reference to the image, so we get an error.
The above three features are all implemented with the Dialog component.